Volume, AudioFiles and more

+++ AudioPlayer/Core/Structures/AudioFile.cs
@@ -7,7 +7,7 @@ namespace AudioPlayer.Core.Structures
     public class AudioFile
     {
         public string Path { get; set; } = System.IO.Path.Combine(Paths.Configs, "Audios", "test.mp3");
-        public int Volume { get; set; } = 100;
+        public float Volume { get; set; } = 100;
 
         public void Play(bool loop = false, bool automatic = false)
         {
